update benchmark script
